requirements.txt
asgiref==3.3.1
Django==3.0.11
django-taggit==1.3.0
Markdown==3.2.1
psycopg2-binary==2.8.4
pytz==2020.4
sqlparse==0.4.1
